make sonar happy again
