+++ app/components/userFeedback/BannerState.ts
@@ -1,0 +1,5 @@
+export enum BannerState {
+  ShowRating = "showRating",
+  ShowFeedback = "showFeedback",
+  FeedbackGiven = "feedbackGiven",
+}
