+++ app/routes/action.send-feedback.ts
@@ -2,7 +2,8 @@ import { parseFormData, validationError } from "@rvf/react-router";
 import type { ActionFunctionArgs } from "react-router";
 import { redirect } from "react-router";
 import { type z } from "zod";
-import { BannerState, USER_FEEDBACK_ID } from "~/components/userFeedback";
+import { USER_FEEDBACK_ID } from "~/components/userFeedback";
+import { BannerState } from "~/components/userFeedback/BannerState";
 import { feedbackSchema } from "~/components/userFeedback/FeedbackFormBox";
 import { userRatingFieldname } from "~/components/userFeedback/RatingBox";
 import { flowIdFromPathname } from "~/domains/flowIds";
