+++ app/routes/action.send-rating.ts
@@ -2,7 +2,8 @@ import { parseFormData, validationError } from "@rvf/react-router";
 import type { ActionFunctionArgs, Session, SessionData } from "react-router";
 import { data, redirect } from "react-router";
 import { z } from "zod";
-import { BannerState, USER_FEEDBACK_ID } from "~/components/userFeedback";
+import { USER_FEEDBACK_ID } from "~/components/userFeedback";
+import { BannerState } from "~/components/userFeedback/BannerState";
 import { userRatingFieldname } from "~/components/userFeedback/RatingBox";
 import { flowIdFromPathname } from "~/domains/flowIds";
 import { sendCustomAnalyticsEvent } from "~/services/analytics/customEvent";
