+++ app/services/feedback/__test__/updateBannerState.test.ts
@@ -1,6 +1,6 @@
 import { createSession, type Session } from "react-router";
 import { describe, test, expect } from "vitest";
-import { BannerState } from "~/components/userFeedback";
+import { BannerState } from "~/components/userFeedback/BannerState";
 import { bannerStateName } from "../getFeedbackData";
 import { updateBannerState } from "../updateBannerState";
