+++ app/services/feedback/getFeedbackData.ts
@@ -1,5 +1,5 @@
 import { type Session } from "react-router";
-import { BannerState } from "~/components/userFeedback";
+import { BannerState } from "~/components/userFeedback/BannerState";
 import { type FeedbackType } from "~/components/userFeedback/FeedbackType";
 
 export const bannerStateName = "bannerState";
