+++ app/services/feedback/updateBannerState.ts
@@ -1,5 +1,5 @@
 import type { Session, SessionData } from "react-router";
-import type { BannerState } from "~/components/userFeedback";
+import { type BannerState } from "~/components/userFeedback/BannerState";
 import { bannerStateName } from "./getFeedbackData";
 
 export function updateBannerState(
