+++ stories/FilesUpload.stories.tsx
@@ -2,11 +2,11 @@ import { faker } from "@faker-js/faker";
 import { ActionFunctionArgs } from "react-router";
 import type { Meta, StoryObj } from "@storybook/react";
 import FilesUpload from "~/components/filesUpload/FilesUpload";
-import { splitFieldName } from "~/services/upload/fileUploadHelpers";
 import { TranslationContext } from "~/services/translations/translationsContext";
 import { PDFFileMetadata, TEN_MB_IN_BYTES } from "~/util/file/pdfFileSchema";
 import { RFCFormerProvider } from ".storybook/RFCFormerProvider";
 import { reactRouterContext } from ".storybook/reactRouterContext";
+import { splitFieldName } from "~/services/upload/splitFieldName";
 
 const meta = {
   title: "Component/FilesUpload",
